+++ src/Model/entities/Manager.java
@@ -1,7 +1,7 @@
 package model.entities;
 
 public class Manager extends Worker{
-    public Manager(String name, String dateOfBirth, int id, double hourlySalary, double numHourMonth, String jobType, String password) {
+    public Manager(String name, String dateOfBirth, int id, int hourlySalary, int numHourMonth, String jobType, String password) {
         super(name, dateOfBirth, id, hourlySalary, numHourMonth, jobType, password);
     }
