+++ src/Model/entities/Pants.java
@@ -10,7 +10,7 @@ public class Pants extends Item {
         this.pantsType = pantsType;
     }
 
-    public Pants(String color, String brand, String gender, String type, double price, int size, int currentStock, int baseStock, int itemId, String pantsType) {
+    public Pants(String color, String brand, String gender, String type, int price, int size, int currentStock, int baseStock, int itemId, String pantsType) {
         super(color, brand, gender, type, price, size, currentStock, baseStock, itemId);
         this.pantsType = pantsType;
     }
