+++ src/Model/entities/Shirt.java
@@ -10,7 +10,7 @@ public class Shirt extends Item {
         this.shirtType = shirtType;
     }
 
-    public Shirt(String color, String brand, String gender, String type, double price, int size, int currentStock, int baseStock, int itemId, String shirtType) {
+    public Shirt(String color, String brand, String gender, String type, int price, int size, int currentStock, int baseStock, int itemId, String shirtType) {
         super(color, brand, gender, type, price, size, currentStock, baseStock, itemId);
         this.shirtType = shirtType;
     }
