+++ src/Model/entities/Shoe.java
@@ -15,7 +15,7 @@ public class Shoe extends Item{
         this.drawstringColor = drawstringColor;
     }
 
-    public Shoe(String color, String brand, String gender, String type, double price, int size, int currentStock, int baseStock, int itemId, String drawstringColor) {
+    public Shoe(String color, String brand, String gender, String type, int price, int size, int currentStock, int baseStock, int itemId, String drawstringColor) {
         super(color, brand, gender, type, price, size, currentStock, baseStock, itemId);
         this.drawstringColor = drawstringColor;
     }
